Integrated infusionsoft:
Done this. This is how I did it:

- The iSDK is housed under libraries/infusionsoft

- There is a simple class defined in libraries called Infusionsoft

- This class connects to the app when instantiated, and also makes all Infusionsoft methods available

- Finally, I created a model (inf_model) in models folder that extends this library.

- We can create methods in this library that, in turn, call infusionsoft methods and call them like this:

$this->load->model('inf_model');	//Alkways load the model somewhere

$result = $this->inf_model->infusion_get('Contact', $id, array('FirstName'));	//object is 'inf_model'

This si prb not the most effective way of doing it (for example I am sure you are able to write this so that the iSDK class is accessible directly, but that's beyond the scope of this app at this time. Pus, by creating an abstracted layer between the methods written in this app, and the functions of the external CRM, we can swap out the CRM at a alter date by simply writing a new inf_library.